requirements.txt
pytorch
jupyter
ipdb
pandas
scikit-learn
choix
lime